- move auth storage logic to usecase\interactor
