+++ internal/usecase/interactor/auth.go
@@ -1,4 +1,4 @@
-package appauth
+package interactor
 
 import (
 	"context"
